Refactoring logger benchmark
