+++ src/backend/benchmark/Makefile.am
@@ -10,9 +10,9 @@
  
 hyadapt_SOURCES =  \
 					backend/benchmark/hyadapt/hyadapt.cpp \
-                    backend/benchmark/hyadapt/configuration.cpp \
-                    backend/benchmark/hyadapt/workload.cpp \
-                    backend/benchmark/hyadapt/loader.cpp
+                    backend/benchmark/hyadapt/hyadapt_configuration.cpp \
+                    backend/benchmark/hyadapt/hyadapt_workload.cpp \
+                    backend/benchmark/hyadapt/hyadapt_loader.cpp
 
 hyadapt_LDFLAGS =
 hyadapt_CPPFLAGS = -I. -I$(top_srcdir)/src -I.. $(postgres_common_INCLUDES) $(AM_CPPFLAGS)  \
@@ -21,3 +21,19 @@ hyadapt_CPPFLAGS = -I. -I$(top_srcdir)/src -I.. $(postgres_common_INCLUDES) $(AM
  
 hyadapt_LDADD = libpelotonpg.la libpeloton.la -lpthread
 
+######################################################################
+# LOGGER
+######################################################################
+ 
+logger_SOURCES =  \
+					backend/benchmark/logger/logger.cpp \
+                    backend/benchmark/logger/logger_configuration.cpp \
+                    backend/benchmark/logger/logger_workload.cpp \
+                    backend/benchmark/logger/logger_loader.cpp
+
+logger_LDFLAGS =
+logger_CPPFLAGS = -I. -I$(top_srcdir)/src -I.. $(postgres_common_INCLUDES) $(AM_CPPFLAGS)  \
+				   $(third_party_INCLUDES) \
+				   -I$(srcdir)/backend/benchmark
+ 
+logger_LDADD = libpelotonpg.la libpeloton.la -lpthread
