+++ src/backend/benchmark/hyadapt/hyadapt.cpp
@@ -14,8 +14,8 @@
 #include <fstream>
 
 #include "backend/benchmark/hyadapt/hyadapt.h"
-#include "backend/benchmark/hyadapt/configuration.h"
-#include "backend/benchmark/hyadapt/workload.h"
+#include "backend/benchmark/hyadapt/hyadapt_configuration.h"
+#include "hyadapt_workload.h"
 
 namespace peloton {
 namespace benchmark {
