+++ src/backend/benchmark/hyadapt/hyadapt_configuration.cpp
@@ -10,10 +10,11 @@
 //
 //===----------------------------------------------------------------------===//
 
+#include "backend/benchmark/hyadapt/hyadapt_configuration.h"
+
 #include <iomanip>
 #include <algorithm>
 
-#include "backend/benchmark/hyadapt/configuration.h"
 
 namespace peloton {
 namespace benchmark {
