+++ src/backend/benchmark/hyadapt/hyadapt_loader.h
@@ -2,9 +2,9 @@
 //
 //                         PelotonDB
 //
-// hyadapt.cpp
+// workload.h
 //
-// Identification: benchmark/hyadapt/hyadapt.h
+// Identification: benchmark/hyadapt/workload.h
 //
 // Copyright (c) 2015, Carnegie Mellon University Database Group
 //
@@ -12,13 +12,21 @@
 
 #pragma once
 
-#include "hyadapt_configuration.h"
+#include "backend/benchmark/hyadapt/hyadapt_configuration.h"
 
 namespace peloton {
 namespace benchmark {
 namespace hyadapt {
 
 extern configuration state;
+
+extern storage::DataTable *hyadapt_table;
+
+void CreateTable();
+
+void LoadTable();
+
+void CreateAndLoadTable(LayoutType layout_type);
 
 }  // namespace hyadapt
 }  // namespace benchmark
