+++ src/backend/benchmark/hyadapt/hyadapt_workload.h
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include "backend/benchmark/hyadapt/configuration.h"
+#include "backend/benchmark/hyadapt/hyadapt_configuration.h"
 
 namespace peloton {
 namespace benchmark {
