+++ tests/logging/logging_tests_util.h
@@ -2,9 +2,6 @@
 
 #include <vector>
 
-#include "backend/common/types.h"
-#include "backend/common/pool.h"
-
 namespace peloton {
 
 namespace catalog {
@@ -18,15 +15,6 @@ class DataTable;
 }
 
 namespace test {
-
-enum LoggingExperimentType {
-  LOGGING_EXPERIMENT_TYPE_INVALID = 0,
-
-  LOGGING_EXPERIMENT_TYPE_ACTIVE = 1,
-  LOGGING_EXPERIMENT_TYPE_RECOVERY = 2,
-  LOGGING_EXPERIMENT_TYPE_STORAGE = 3,
-  LOGGING_EXPERIMENT_TYPE_WAIT = 4
-};
 
 class LoggingTestsUtil {
  public:
@@ -49,36 +37,6 @@ class LoggingTestsUtil {
   //===--------------------------------------------------------------------===//
 
   static void ParseArguments(int argc, char* argv[]);
-
-  class logging_test_configuration {
-   public:
-    // experiment type
-    LoggingExperimentType experiment_type;
-
-    // logging type
-    LoggingType logging_type;
-
-    // # of tuples
-    int tuple_count;
-
-    // # of backends (i.e. backend loggers)
-    int backend_count;
-
-    // # of columns in each tuple
-    oid_t column_count;
-
-    // check if the count matches after recovery
-    bool check_tuple_count;
-
-    // log file dir
-    std::string log_file_dir;
-
-    // size of the pmem file (in MB)
-    size_t data_file_size;
-
-    // frequency with which the logger flushes
-    int64_t wait_timeout;
-  };
 
  private:
   //===--------------------------------------------------------------------===//
@@ -124,9 +82,6 @@ class LoggingTestsUtil {
 
   static void CheckTupleCount(oid_t db_oid, oid_t table_oid, oid_t expected);
 };
-
-// configuration for testing
-extern LoggingTestsUtil::logging_test_configuration state;
 
 }  // End test namespace
 }  // End peloton namespace
